file_size expr cleanup
