+++ src/daft-file/src/lib.rs
@@ -1,5 +1,9 @@
-pub mod functions;
-#[cfg(feature = "python")]
-pub mod python;
+mod file;
+mod functions;
 
 pub use functions::*;
+
+pub use crate::file::DaftFile;
+
+#[cfg(feature = "python")]
+pub mod python;
